rename sub component dir to ..Utility... to point out the use for included components
