+++ factsheet/frontend/src/components/scenarioBundleUtilityComponents/StudyDescriptors.js
